requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_tflow_select=2.1.0=gpu
absl-py=0.9.0=py37_0
asn1crypto=1.3.0=py37_0
astor=0.8.0=py37_0
blas=1.0=mkl
blinker=1.4=py37_0
c-ares=1.15.0=h7b6447c_1001
ca-certificates=2020.1.1=0
cachetools=3.1.1=py_0
cairo=1.14.12=h8948797_3
certifi=2019.11.28=py37_1
cffi=1.14.0=py37h2e261b9_0
chardet=3.0.4=py37_1003
click=7.1.1=py_0
cryptography=2.8=py37h1ba5d50_0
cudatoolkit=10.1.243=h6bb024c_0
cudnn=7.6.5=cuda10.1_0
cupti=10.1.168=0
cycler=0.10.0=py_2
dbus=1.13.12=h746ee38_0
expat=2.2.9=he1b5a44_2
fontconfig=2.13.0=h9420a91_0
freetype=2.9.1=h8a8886c_1
fribidi=1.0.5=h7b6447c_0
gast=0.2.2=py37_0
gettext=0.19.8.1=hc5be6a0_1002
glib=2.56.2=hd408876_0
google-auth=1.11.2=py_0
google-auth-oauthlib=0.4.1=py_2
google-pasta=0.1.8=py_0
graphite2=1.3.13=h23475e2_0
graphviz=2.40.1=h21bd128_2
grpcio=1.27.2=py37hf8bcb03_0
gst-plugins-base=1.14.0=hbbd80ab_1
gstreamer=1.14.0=hb453b48_1
h5py=2.10.0=py37h7918eee_0
harfbuzz=1.8.8=hffaf4a1_0
hdf5=1.10.4=hb1b8bf9_0
icu=58.2=h211956c_0
idna=2.9=py_1
intel-openmp=2020.0=166
joblib=0.14.1=py_0
jpeg=9c=h14c3975_1001
keras-applications=1.0.8=py_0
keras-preprocessing=1.1.0=py_1
kiwisolver=1.1.0=py37h99015e2_1
ld_impl_linux-64=2.33.1=h53a641e_7
libclang=9.0.1=default_hde54327_0
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libiconv=1.15=h516909a_1006
libllvm9=9.0.1=hc9558a2_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.11.4=hd408876_0
libstdcxx-ng=9.1.0=hdf63c60_0
libtiff=4.1.0=h2733197_0
libuuid=1.0.3=h1bed415_2
libxcb=1.13=h14c3975_1002
libxkbcommon=0.10.0=he1b5a44_0
libxml2=2.9.9=hea5a465_1
markdown=3.1.1=py37_0
matplotlib=3.1.3=py37_0
matplotlib-base=3.1.3=py37hef1b27d_0
mkl=2019.5=281
mkl-service=2.3.0=py37he904b0f_0
mkl_fft=1.0.15=py37ha843d7b_0
mkl_random=1.1.0=py37hd6b4f25_0
ncurses=6.2=he6710b0_0
ninja=1.9.0=py37hfd86e86_0
nspr=4.25=he1b5a44_0
nss=3.47=he751ad9_0
numpy=1.18.1=py37h4f9e942_0
numpy-base=1.18.1=py37hde5b4d6_1
oauthlib=3.1.0=py_0
olefile=0.46=py_0
openssl=1.1.1=h7b6447c_0
opt_einsum=3.1.0=py_0
pango=1.42.4=h049681c_0
pcre=8.44=he1b5a44_0
pillow=7.0.0=py37hb39fc2d_0
pip=20.0.2=py37_1
pixman=0.38.0=h7b6447c_0
protobuf=3.11.4=py37he6710b0_0
pthread-stubs=0.4=h14c3975_1001
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.7=py_0
pycparser=2.20=py_0
pydot=1.4.1=py37_0
pyjwt=1.7.1=py37_0
pyopenssl=19.1.0=py37_0
pyparsing=2.4.6=py_0
pyqt=5.9.2=py37h22d08a2_1
pysocks=1.7.1=py37_0
python=3.7.7=h191fe78_0_cpython
python-dateutil=2.8.1=py_0
python_abi=3.7=1_cp37m
pytorch=1.4.0=py3.7_cuda10.1.243_cudnn7.6.3_0
qt=5.9.7=h5867ecd_1
readline=7.0=h7b6447c_5
requests=2.23.0=py37_0
requests-oauthlib=1.3.0=py_0
rsa=4.0=py_0
scikit-learn=0.22.1=py37hd81dba3_0
scipy=1.4.1=py37h0b6359f_0
setuptools=46.1.1=py37_0
sip=4.19.13=py37he6710b0_0
six=1.14.0=py37_0
sqlite=3.31.1=h7b6447c_0
tensorboard=2.1.0=py3_0
tensorflow=2.1.0=gpu_py37h7a4bb67_0
tensorflow-base=2.1.0=gpu_py37h6c5654b_0
tensorflow-estimator=2.1.0=pyhd54b08b_0
tensorflow-gpu=2.1.0=h0d30ee6_0
termcolor=1.1.0=py37_1
tk=8.6.10=hed695b0_0
torchvision=0.5.0=py37_cu101
tornado=6.0.4=py37h8f50634_1
tqdm=4.43.0=py_0
urllib3=1.25.8=py37_0
werkzeug=1.0.0=py_0
wheel=0.34.2=py37_0
wrapt=1.12.1=py37h7b6447c_1
xorg-libxau=1.0.9=h14c3975_0
xorg-libxdmcp=1.1.3=h516909a_0
xz=5.2.4=h14c3975_4
zlib=1.2.11=h7b6447c_3
zstd=1.3.7=h0b5b093_0
